pyproject.toml
[project]
name = "template-alpine"
version = "0.1.0"
description = "Template for Alpine.js + Quart web projects"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hypercorn>=0.17.0",
    "psycopg>=3.3.3",
    "python-dotenv>=1.2.2",
    "quart-cors>=0.8.0",
    "quart>=0.20.0",
    "sqlalchemy>=2.0.48",
]
